+++ HISTORY.md
@@ -4,7 +4,7 @@
 
 ### API
 
-- Display dependencies on click (#25)
+- Display dependencies on click (#25, #26)
 - Migrate from graph objects to Dash (#20)
 - Enable arrow keys for time travel (#20)
 - Remove "null" text in empty cells (#21)